+++ src/Raven.CodeAnalysis/BoundTree/BoundYieldExpression.cs
@@ -8,15 +8,18 @@ internal sealed partial class BoundYieldExpression : BoundExpression
         BoundExpression expression,
         ITypeSymbol elementType,
         IteratorMethodKind iteratorKind,
-        ITypeSymbol type)
+        ITypeSymbol type, ForIterationInfo? iteration = null)
         : base(type, symbol: null, BoundExpressionReason.None)
     {
         Expression = expression;
         ElementType = elementType;
         IteratorKind = iteratorKind;
+        Iteration = iteration;
     }
 
     public BoundExpression Expression { get; }
+
+    public ForIterationInfo? Iteration { get; }
 
     public ITypeSymbol ElementType { get; }